+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/JsonoidParams.scala
@@ -3,20 +3,25 @@ package edu.rit.cs.mmior.jsonoid.discovery
 final case class JsonoidParams(
     val additionalProperties: Boolean = false,
     val er: EquivalenceRelation = EquivalenceRelations.KindEquivalenceRelation,
+    val formatThreshold: Float = 1.0f,
     val maxExamples: Int = 100
 ) {
   def withAdditionalProperties(
       newAdditionalProperties: Boolean
   ): JsonoidParams = {
-    JsonoidParams(newAdditionalProperties, er, maxExamples)
+    JsonoidParams(newAdditionalProperties, er, formatThreshold, maxExamples)
   }
 
   def withER(newER: EquivalenceRelation): JsonoidParams = {
-    JsonoidParams(additionalProperties, newER, maxExamples)
+    JsonoidParams(additionalProperties, newER, formatThreshold, maxExamples)
+  }
+
+  def withFormatThreshold(newFormatThreshold: Float): JsonoidParams = {
+    JsonoidParams(additionalProperties, er, newFormatThreshold, maxExamples)
   }
 
   def withMaxExamples(newMaxExamples: Int): JsonoidParams = {
-    JsonoidParams(additionalProperties, er, newMaxExamples)
+    JsonoidParams(additionalProperties, er, formatThreshold, newMaxExamples)
   }
 }